+++ src/ai/flows/validate-visual-standard.ts
@@ -4,37 +4,10 @@
  * @fileOverview A flow for validating a user's image submission against a PDF visual standard.
  *
  * - validateVisualStandard - A function that orchestrates the validation process.
- * - ValidateVisualStandardInput - The input type for the validation function.
- * - ValidateVisualStandardOutput - The return type for the validation function.
  */
 
 import {ai} from '@/ai/genkit';
-import {z} from 'genkit';
-
-// Define the structure for a failed zone
-const FailedZoneSchema = z.object({
-  zoneId: z.string().describe('The ID of the zone that failed validation.'),
-  reason: z.string().describe('The specific reason for the failure.'),
-});
-
-// Define the input schema for the flow
-export const ValidateVisualStandardInputSchema = z.object({
-  photoDataUri: z
-    .string()
-    .describe(
-      "A photo submitted by the user, as a data URI that must include a MIME type and use Base64 encoding. Expected format: 'data:<mimetype>;base64,<encoded_data>'."
-    ),
-  standardPdfUrl: z.string().url().describe('The URL of the standard PDF to validate against.'),
-});
-export type ValidateVisualStandardInput = z.infer<typeof ValidateVisualStandardInputSchema>;
-
-// Define the output schema for the flow
-export const ValidateVisualStandardOutputSchema = z.object({
-  status: z.enum(['OK', 'NOT_OK']).describe('The overall validation status.'),
-  failureReason: z.string().optional().describe('A summary of why the submission failed.'),
-  failedZones: z.array(FailedZoneSchema).optional().describe('A list of zones that failed validation.'),
-});
-export type ValidateVisualStandardOutput = z.infer<typeof ValidateVisualStandardOutputSchema>;
+import { ValidateVisualStandardInput, ValidateVisualStandardInputSchema, ValidateVisualStandardOutput, ValidateVisualStandardOutputSchema } from './validate-visual-standard-types';
 
 
 /**
